Services/Plugins/pi_metrics/requirements.txt
paho-mqtt
psutil
python-dotenv